install: update the operator without removing it first
it could still be removed by passing "uninstall" as argument

+++ elemental-install.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+ARGS="$1"
 EO_NAME="elemental-operator"
 EO_SRC="/home/fgiudici/go/src/github.com/rancher/elemental-operator"
 EO_REPO="quay.io/fgiudici/elemental-operator"
@@ -45,15 +46,15 @@ build_docker() {
 }
 
 install_helm_chart() {
+    echo "Install $CHART_NAME" 
     pushd ${EO_SRC}
-    kubectl create namespace ${EO_NS}
     set -x
-    helm install -n ${EO_NS} --set image.repository=${EO_REPO} ${EO_NAME} ${CHART_NAME} --set debug=true
+    helm upgrade --create-namespace -n ${EO_NS} --install ${EO_NAME} ${CHART_NAME} --set image.repository=${EO_REPO} --set debug=true
     set +x
     popd
 }
 
-if is_operator_installed; then
+if [ "$ARGS" = "uninstall" ] && is_operator_installed; then
     echo "Removing already installed operator"
     helm uninstall -n cattle-elemental-system elemental-operator
